SupremeAI-Chatbot/requirements.txt
botbuilder-core
botbuilder-schema
botbuilder-dialogs
botbuilder-ai
flask 
azure-search-documents 
azure-cosmos 
azure-ai-language-conversations